Add ability to join a game.
This is a workaround for the problem that players sometimes get kicked out of a game when a new player joins (which should really be solved with opti-locking on the game). It also allows the user to open a game URL and join it from there.

Also some minor fixes to other parts of the app and added a basic GameBuilder class for creating test games.